fix: review corrections across PR #295-303 stack
- pkg/session/postgres.go: atomic INSERT for AppendTurn (TOCTOU race
  under concurrent writers); errors.Is for sql.ErrNoRows
- pkg/claude/options.go: fix double --resume in persistentRestartArgs;
  correct RetryMaxAttempts zero-value doc
- pkg/claude/persistent.go: set StopReasonError on non-budget errors;
  remove dead maxAttempts>0 guard
- pkg/a2a/outbound.go: RWMutex + double-checked locking to avoid holding
  lock during network I/O; SSRF scheme and userinfo validation
- pkg/a2a/card.go: rename bare AGENT_* env vars to KKAUS_AGENT_* to
  avoid silent collisions in multi-container pods
- pkg/a2a/executor.go: guard AppendTurn behind sessionID != "" to prevent
  NOT NULL violation in chat mode
- pkg/config/config.go: trim whitespace in envOverrideCSV so
  KKAUS_A2A_ALLOWED_HOSTS values with spaces are matched correctly
- pkg/kagentapi/client.go: cache agentRef at construction; PathEscape
  sessionID in URL; drain response body before close; avoid url pkg shadow
- pkg/memory/client.go: Chunk.Score float32 -> float64 (matches API)
- pkg/memory/kagent.go: X-User-ID -> X-User-Id (match kagentapi); remove
  float32 cast; update doc comment
- helm/klaus/templates/deployment.yaml: render a2a.* values as env vars
  (targets, allowDynamic, allowedHosts were dead without this)
- helm/values.yaml: correct a2a value types and alignment
- docs/reference/helm-values.md: fix postgresSecretKey default ("uri")
- CHANGELOG.md: add Unreleased entries for the full A2A/session/memory
  stack; note automountServiceAccountToken change

+++ pkg/memory/client.go
@@ -40,7 +40,7 @@ type Chunk struct {
 	// Content is the recalled text to inject into the system prompt.
 	Content string
 	// Score is the cosine similarity score (0–1). Higher is more relevant.
-	Score float32
+	Score float64
 }
 
 // Client retrieves and stores conversation memory for a context.
